requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
httpx>=0.26.0
parsel>=1.9.0
readability-lxml>=0.8.1
textstat>=0.7.3
lxml>=5.1.0
pydantic>=2.6.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0 